Add simple agent script for testing envs. Fix bugs/warnings in EdgeGrasp and base envs

+++ src/contact_gym/envs/base.py
@@ -38,8 +38,8 @@ class MujocoBaseEnv(ABC, gym.Env):
         self._renderer = None
         self._home_key_id = self._set_home_key()
         # Setup action space
-        ctrl_low = self.model.actuator_ctrlrange[:, 0]
-        ctrl_high = self.model.actuator_ctrlrange[:, 1]
+        ctrl_low = self.model.actuator_ctrlrange[:, 0].astype(np.float32)
+        ctrl_high = self.model.actuator_ctrlrange[:, 1].astype(np.float32)
         self.action_space = spaces.Box(low=ctrl_low, high=ctrl_high, dtype=np.float32)
 
     # region Core API
